+++ packages/pds/src/api/app/bsky/graph/unsubscribeMuteList.ts
@@ -2,7 +2,7 @@ import { Server } from '../../../../lexicon'
 import AppContext from '../../../../context'
 
 export default function (server: Server, ctx: AppContext) {
-  server.app.bsky.graph.subscribeMuteList({
+  server.app.bsky.graph.unsubscribeMuteList({
     auth: ctx.accessVerifier,
     handler: async ({ auth, input }) => {
       const { list } = input.body
